+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/deprecated/ALPHAvisionAuto.java
@@ -1,14 +1,38 @@
-package org.firstinspires.ftc.teamcode.autonomous;
+/*
+ * Copyright (c) 2024 Phil Malone
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package org.firstinspires.ftc.teamcode.deprecated;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 import org.firstinspires.ftc.teamcode.RobotHardware;
 
-@Autonomous(name= "BUCKET Side Auto", group ="Robot")
-public class BucketSideAuto extends LinearOpMode {
+@Disabled
+@Autonomous(name = "ALPHA auto by Vision", group = "Robot")
+public class ALPHAvisionAuto extends LinearOpMode{
 
     RobotHardware robot = new RobotHardware(this);
     ElapsedTime runtime = new ElapsedTime();
@@ -21,6 +45,21 @@ public class BucketSideAuto extends LinearOpMode {
         robot.init();
         robot.elbowDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
         robot.elbowDrive.setPower(1.0);
+
+        robot.leftFrontDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        robot.leftBackDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        robot.rightFrontDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        robot.rightBackDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+
+        robot.leftFrontDrive.setTargetPosition(0);
+        robot.leftBackDrive.setTargetPosition(0);
+        robot.rightFrontDrive.setTargetPosition(0);
+        robot.rightBackDrive.setTargetPosition(0);
+
+        robot.leftFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+        robot.leftBackDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+        robot.rightFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+        robot.rightBackDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
         // Send a telemetry message to signify the robot waiting; wait for the game to start (driver presses PLAY)
         waitForStart();
         runtime.reset();
@@ -39,7 +78,7 @@ public class BucketSideAuto extends LinearOpMode {
         robot.driveFieldCentric(0.3,0,0);
         robot.elbowDrive.setTargetPosition((int) (robot.ELBOW_PERPENDICULAR - robot.angleConvert(60))); // score/ hook on
         runtime.reset();
-        while (opModeIsActive() && runtime.seconds() < 0.75  ) {
+        while (opModeIsActive() && runtime.seconds() < 0.65  ) {
             telemetry.addData("SCORING", "...");
             telemetry.update();
             robot.setClawPosition(robot.pass,robot.superposition,robot.superposition);
@@ -73,16 +112,109 @@ public class BucketSideAuto extends LinearOpMode {
         robot.driveFieldCentric(0,0,0);
 
         sleep(500);
-
-        robot.driveFieldCentric(0,-1.0,0);
+        /*
+        robot.setDrivePower(0.43,-1.0,-1.0,0.43); // strafe to OZ
+        robot.setClawPosition(robot.disable,0,robot.pass);
         runtime.reset();
-        while (opModeIsActive() && runtime.seconds() < 0.7) {
+        while (opModeIsActive() && runtime.seconds() < 0.6) {
             telemetry.addData("STRAFING", "...");
             telemetry.update();
         }
         robot.driveFieldCentric(0,0,0);
 
-        // score done!
+        sleep(3000);
+
+        robot.driveFieldCentric(0,0,-1.0);
+        runtime.reset();
+        while (opModeIsActive() && runtime.seconds() < 1.5) {
+            telemetry.addData("TURNING", "...");
+            telemetry.update();
+        }
+        robot.driveFieldCentric(0,0,0);
+
+        sleep(10000); // TEST THIS BEFORE MOVING ON  ^^^^^^^^^
+
+        colorDetector.visionInit("BLUE",true, colorDetector.colorLocator, -1,0.6255,0.5,-0.625); // scan for the specimen
+        robot.driveFieldCentric(0,-0.625,0); // USE THIS SAME SPEED FOR secondsToScan
+        runtime.reset();
+        while (opModeIsActive() && runtime.seconds() < 1) {
+            robot.calibrateClaw(robot.ELBOW_PARALLEL);
+            secondsToScan = runtime.seconds();
+            colorDetector.activeDetector(new Point(480,810), new Point(1440,270), "PRIMARY");
+            if (!colorDetector.blobS.isEmpty()) {
+                robot.driveFieldCentric(0,0,0);
+                secondsToScan = runtime.seconds();
+                break;
+            }
+        }
+
+        sleep(300);
+
+        robot.driveFieldCentric(0,0.625,0); // REVERSE THE DRIVING FROM ABOVE ^^^
+        runtime.reset();
+        while (opModeIsActive() && runtime.seconds() < secondsToScan) {
+            telemetry.addData("MOVING BACK", "...");
+            telemetry.update();
+        }
+
+        robot.driveFieldCentric(0.625,0,0); // move up to grab specimen
+        runtime.reset();
+        while (opModeIsActive() && runtime.seconds() < 0.2) {
+            telemetry.addData("DOCKING", "...");
+            telemetry.update();
+        }
+        robot.driveFieldCentric(0,0,0);
+
+        sleep(200);
+
+        robot.setClawPosition(robot.enable,0,robot.enable); // grab specimen
+        robot.elbowDrive.setTargetPosition((int) (robot.ELBOW_PERPENDICULAR + robot.angleConvert(15))); // lift arm
+        robot.setDrivePower(1.0,-1.0,-0.43,0.43); // strafe back to rung
+        runtime.reset();
+        while (opModeIsActive() && runtime.seconds() < 0.6) {
+            telemetry.addData("STRAFING", "...");
+            telemetry.update();
+        }
+        robot.driveFieldCentric(0,0,0);
+
+        sleep(200);
+
+        robot.driveFieldCentric(0.6,0,0); // move up/reverse of back away
+        runtime.reset();
+        while (opModeIsActive() && runtime.seconds() < 0.01) {
+            telemetry.addData("MOVING UP", "...");
+            telemetry.update();
+        }
+
+        robot.driveFieldCentric(0.3,0,0); // hook onto rung/score
+        robot.elbowDrive.setTargetPosition((int) (robot.ELBOW_PERPENDICULAR + robot.angleConvert(60)));
+        runtime.reset();
+        while (opModeIsActive() && runtime.seconds() < 0.2  ) {
+            telemetry.addData("SCORING", "...");
+            telemetry.update();
+        }
+        robot.driveFieldCentric(0,0,0);
+
+        while (robot.elbowDrive.isBusy()) { // check for when to let go of specimen
+            robot.setClawPosition(robot.pass,0,robot.superposition);
+
+            if (robot.elbowDrive.getCurrentPosition() == (int) (robot.COUNTS_PER_DEGREE * 180) ) {// test and edit this #
+                break;
+            }
+        }
+        robot.setClawPosition(robot.disable,0,robot.pass); // let go of specimen
+        telemetry.addData("GET ELBOW ANGLE", robot.elbowDrive.getCurrentPosition()/robot.COUNTS_PER_DEGREE );
+        telemetry.update();
+
+        */
+        // bring arm down and perpendicularize claw before this, also assume that this is after your FIRST/PRELOADED specimen
+        robot.driveFieldCentric(0,1.0,0);
+        runtime.reset();
+        while (opModeIsActive() && runtime.seconds() < 0.55) {
+            telemetry.addData("STRAFING", "...");
+            telemetry.update();
+        }
+        robot.driveFieldCentric(0,0,0);
 
         sleep(200);
 
@@ -96,9 +228,9 @@ public class BucketSideAuto extends LinearOpMode {
 
         sleep(200);
 
-        robot.driveFieldCentric(0,-1.0,0);
+        robot.driveFieldCentric(0,1.0,0.);
         runtime.reset();
-        while (opModeIsActive() && runtime.seconds() < 0.4) {
+        while (opModeIsActive() && runtime.seconds() < 0.42){
             telemetry.addData("STRAFING", "...");
             telemetry.update();
         }
@@ -162,11 +294,9 @@ public class BucketSideAuto extends LinearOpMode {
         robot.elbowDrive.setTargetPosition((int) (robot.ELBOW_COLLAPSED));
         sleep(200);
 
-        // strafe to sub zone
-
         robot.driveFieldCentric(0,1.0,0);
         runtime.reset();
-        while (opModeIsActive() && runtime.seconds() < 0.4) {
+        while (opModeIsActive() && runtime.seconds() < 0.5 ) {
             telemetry.addData("STRAFING", "...");
             telemetry.update();
         }
@@ -174,12 +304,48 @@ public class BucketSideAuto extends LinearOpMode {
 
         sleep(200);
 
-        robot.elbowDrive.setTargetPosition((int) robot.ELBOW_COLLAPSED);
+        robot.heading = robot.imu.getRobotYawPitchRollAngles().getYaw();
+
+        if (opModeIsActive() && robot.heading !=0) {
+            robot.driveFieldCentric(0, 0, robot.turnDirection(0, true)); // ADJUST ANGLE
+        }
+        robot.heading = robot.imu.getRobotYawPitchRollAngles().getYaw();
+        while (opModeIsActive() && robot.heading != 0){
+            heading = Math.round(robot.imu.getRobotYawPitchRollAngles().getYaw(AngleUnit.DEGREES)/1.5 ) * 1.5;
+            if (heading == -3) { // go 20 under your target
+                robot.driveFieldCentric(0,0,0);
+                telemetry.addData("AT 180 DEG", "");
+                break;
+            }
+            telemetry.addData("HEADING:", heading);
+            telemetry.update();
+        }
+        sleep(200);
+
+        robot.driveFieldCentric(-1,0,0);
+        runtime.reset();
+        while (opModeIsActive() && runtime.seconds() < 0.775) {
+            telemetry.addData("DRIVING BACK", "...");
+            telemetry.update();
+        }
+        robot.driveFieldCentric(0,0,0);
+
+        sleep(200);
+
+        robot.driveFieldCentric(-0.2,0,0);
+        runtime.reset();
+        while (opModeIsActive() && runtime.seconds() < 0.2) {
+            telemetry.addData("DRIVING FORWARD", "...");
+            telemetry.update();
+        }
+        robot.driveFieldCentric(0,0,0);
+
+        sleep(200);
 
         robot.heading = robot.imu.getRobotYawPitchRollAngles().getYaw();
 
         if (opModeIsActive() && robot.heading !=0) {
-            robot.driveFieldCentric(0, 0, robot.turnDirection(0, true) ); // ADJUST ANGLE
+            robot.driveFieldCentric(0, 0, robot.turnDirection(0, true) * 0.3); // ADJUST ANGLE
         }
         robot.heading = robot.imu.getRobotYawPitchRollAngles().getYaw();
         while (opModeIsActive() && robot.heading != 0){
@@ -192,55 +358,10 @@ public class BucketSideAuto extends LinearOpMode {
             telemetry.addData("HEADING:", heading);
             telemetry.update();
         }
-
-        sleep(5000);
-
-        robot.driveFieldCentric(-1.0,0,0);
-        runtime.reset();
-        while (opModeIsActive() && runtime.seconds() < 0.7) {
-            telemetry.addData("DRIVING BACK", "...");
-            telemetry.update();
-        }
-        robot.driveFieldCentric(0,0,0);
-
-        robot.heading = robot.imu.getRobotYawPitchRollAngles().getYaw();
-
-        if (opModeIsActive() && robot.heading !=0) {
-            robot.driveFieldCentric(0, 0, robot.turnDirection(0, true)); // ADJUST ANGLE
-        }
-        robot.heading = robot.imu.getRobotYawPitchRollAngles().getYaw();
-        while (opModeIsActive() && robot.heading != 0){
-            heading = Math.round(robot.imu.getRobotYawPitchRollAngles().getYaw(AngleUnit.DEGREES)/10) * 10;
-            if (heading == 0) { // go 20 under your target
-                robot.driveFieldCentric(0,0,0);
-                telemetry.addData("AT 180 DEG", "");
-                break;
-            }
-            telemetry.addData("HEADING:", heading);
-            telemetry.update();
-        }
-
         sleep(200);
-
-        robot.driveFieldCentric(1.0,0,0);
-        runtime.reset();
-        while (opModeIsActive() && runtime.seconds() < 0.7) {
-            telemetry.addData("DRIVING", "...");
-            telemetry.update();
-        }
-        robot.driveFieldCentric(0,0,0);
-        robot.elbowDrive.setTargetPosition((int) (robot.ELBOW_COLLAPSED));
-
-        sleep(200);
-
-        robot.driveFieldCentric(0,1.0,0);
-        runtime.reset();
-        while (opModeIsActive() && runtime.seconds() < 0.8) {
-            telemetry.addData("DRIVING", "...");
-            telemetry.update();
-        }
-        robot.driveFieldCentric(0,0,0);
+        // scoring time!
 
         telemetry.addData("DONE","!!");
+
     }
 }
